Rename all ableton components to avoid future name conflicts
