+++ src/app/levelone/(pages)/ableton/_components_/AbletonBodyContent.tsx
@@ -2,7 +2,7 @@ import Image from 'next/image';
 import { PropsWithChildren } from 'react';
 import { studio001 } from './images';
 
-export const Section = (props: PropsWithChildren) => {
+export const AbletonSection = (props: PropsWithChildren) => {
 	return (
 		<section className='flex flex-col gap-4 w-4/5 lg:max-w-prose'>
 			{props.children}
@@ -10,7 +10,7 @@ export const Section = (props: PropsWithChildren) => {
 	);
 };
 
-export const Hero = () => {
+export const AbletonHero = () => {
 	return (
 		<div className='flex flex-col items-center w-4/5 lg:w-full max-w-screen-2xl overflow-hidden'>
 			<h1 className='block absolute z-10 top-1/2 -translate-y-1/2 text-9xl text-red-500'>
